Algorithm: Apply Gaussian Blur using Distance Map
This is my implementation of an algorithm that applies a select Gaussian Blur to an image using a distance map. 

Elijah and I were both assigned the task of implementing a blurring algorithm. This is my implementation that uses the two-pass technique. Elijah and I both created our own implementations independently. Our finalized blurring algorithm used is based on my implementation.